Utopia-925 Refactor NextSteps
